Add Colors and shorten prompt
I added a few colors to make it seem more intrusive.

I also changed a few colors to have a pattern.

I then proceeded to change the word `at` to `@` that makes it look more modern. I also updated the Symbol to $ as > is usually referenced as a comparison and > is usually used to make arrows in themes.